fix data type for max bitrate - int type allows max bitrate of 2147m and this is too low for 10GB network - data type changed to unsigned long

+++ rate-limit.c
@@ -10,7 +10,7 @@
 struct rate_limit {
     long long date;
     long long realDate;
-    int bitrate;
+    unsigned long bitrate;
     int queueSize;
 };
